added this.toggle() and this.isPlaying
I just needed some state to indicate if the simulation is running or
not. During the initialization phase I setup the play button to the
starting state (show the play symbol) since I assume runway-browser does
not init to a running sim, furthermore I call toggle(cb) on every click
and check the sim state within toggle to start or stop the sim. I call
the callback with the playing state (true is running, false is stopped) to
take care of updating the UI.